+++ ibis/backends/sql/compilers/snowflake.py
@@ -1,6 +1,10 @@
 from __future__ import annotations
 
+import inspect
 import itertools
+import platform
+import sys
+import textwrap
 from functools import partial
 
 import sqlglot as sg
@@ -94,6 +98,92 @@ class SnowflakeCompiler(SQLGlotCompiler):
     def __init__(self):
         super().__init__()
         self.f = SnowflakeFuncGen()
+
+    def _compile_python_udf(self, udf_node: ops.ScalarUDF) -> str:
+        return """\
+{preamble}
+HANDLER = '{func_name}'
+AS $$
+from __future__ import annotations
+
+from typing import *
+
+{source}
+$$""".format(**self._get_udf_source(udf_node))
+
+    def _compile_pandas_udf(self, udf_node: ops.ScalarUDF) -> str:
+        template = """\
+{preamble}
+HANDLER = 'wrapper'
+AS $$
+from __future__ import annotations
+
+from typing import *
+
+import _snowflake
+import pandas as pd
+
+{source}
+
+@_snowflake.vectorized(input=pd.DataFrame)
+def wrapper(df):
+    return {func_name}(*(col for _, col in df.items()))
+$$"""
+        return template.format(**self._get_udf_source(udf_node))
+
+    def _get_udf_source(self, udf_node: ops.ScalarUDF):
+        import ibis
+
+        name = type(udf_node).__name__
+        signature = ", ".join(
+            f"{name} {self.type_mapper.to_string(arg.dtype)}"
+            for name, arg in zip(udf_node.argnames, udf_node.args)
+        )
+        return_type = SnowflakeType.to_string(udf_node.dtype)
+        lines, _ = inspect.getsourcelines(udf_node.__func__)
+        source = textwrap.dedent(
+            "".join(
+                itertools.dropwhile(
+                    lambda line: not line.lstrip().startswith("def "), lines
+                )
+            )
+        ).strip()
+
+        config = udf_node.__config__
+
+        preamble_lines = [*self._UDF_PREAMBLE_LINES]
+
+        if imports := config.get("imports"):
+            preamble_lines.append(f"IMPORTS = ({', '.join(map(repr, imports))})")
+
+        packages = "({})".format(
+            ", ".join(map(repr, ("pandas", *config.get("packages", ()))))
+        )
+        preamble_lines.append(f"PACKAGES = {packages}")
+
+        return dict(
+            source=source,
+            name=name,
+            func_name=udf_node.__func_name__,
+            preamble="\n".join(preamble_lines).format(
+                name=name,
+                signature=signature,
+                return_type=return_type,
+                comment=f"Generated by ibis {ibis.__version__} using Python {platform.python_version()}",
+                version=".".join(
+                    map(str, min(sys.version_info[:2], self._latest_udf_python_version))
+                ),
+            ),
+        )
+
+    _UDF_PREAMBLE_LINES = (
+        "CREATE OR REPLACE TEMPORARY FUNCTION {name}({signature})",
+        "RETURNS {return_type}",
+        "LANGUAGE PYTHON",
+        "IMMUTABLE",
+        "RUNTIME_VERSION = '{version}'",
+        "COMMENT = '{comment}'",
+    )
 
     @staticmethod
     def _minimize_spec(start, end, spec):
